refactor(parse): Move parseWordDictEntry to its own file
